requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
